im - css - tidy broke HTML
Apparently, you can't have a div in a label.  I was relying on that because it
allows me to put a <input> in a <label> so that you don't have to use for/id to
link the input/label together.  The issue is that if the input is in the label,
you can no longer use CSS input:checked to refer to a div outside of the label.

My solution was to put the div in the <label>.  And it works in Chrome.  But
its not legal HTML, so tidy breaks it.